+++ src/connectWindow/images/index.js
@@ -5,6 +5,8 @@ import refresh from './refreshIcon';
 import playStoreButton from './button-google-play-color';
 import appStoreButton from './button-app-store';
 import camera from './camera';
+import closeIconWhite from './closeIconWhite'
+import closeIconBlack from './closeIconBlack'
 
 export {
   logoMEW,
@@ -13,5 +15,7 @@ export {
   spaceman,
   playStoreButton,
   appStoreButton,
-  camera
+  camera,
+  closeIconWhite,
+  closeIconBlack
 };