+++ src/connectWindow/popupWindowDesign.js
@@ -166,7 +166,8 @@ const cssStyles = `
       
       .bottom-link {
       text-decoration: none;
-       color: rgba(0, 0, 0, 0.55);
+       color: rgba(5, 192, 165);
+       cursor: pointer;
       }
 
       .refreshIcon {
@@ -195,6 +196,18 @@ const cssStyles = `
           bottom: -2px;
           height: 14px;
           width: 14px;
+        }
+        
+        #google-link:hover {
+        cursor: pointer;
+        }
+        
+        #apple-link:hover {
+        cursor: pointer;
+        }
+        
+        .warn-color {
+        color: orange;
         }
     `;
 
@@ -242,6 +255,7 @@ const htmlDesign = (refresh, image, playStore, appStore, camera) => {
           <div class="center">
             <div class="right">
               <p class="get-text">Don't have MEW wallet app?</p>
+              <p id="popupsBlocked" class="warn-color hidden">Please disable popup blockers to open app or play store link</p>
                <img
                   id="apple-link"
                   class="left-img"
@@ -263,8 +277,8 @@ const htmlDesign = (refresh, image, playStore, appStore, camera) => {
 
         </div>
         <div class="bottom">
-          Powered by <a href="#" class="bottom-link">MEWconnect protocol</a> <br />
-          brought to you by MyEtherWallet
+          Powered by <span id="proto-link" class="bottom-link">MEWconnect protocol</span> <br />
+          brought to you by <span id="mew-link" class="bottom-link">MyEtherWallet</span>
         </div>
       </div>
     </div>
@@ -274,7 +288,9 @@ const htmlDesign = (refresh, image, playStore, appStore, camera) => {
   const refreshButton = window.document.getElementById("refresh");
   const appleLinkButton = window.document.getElementById("apple-link");
   const googleLinkButton = window.document.getElementById("google-link");
-  
+  const mewWebLink = window.document.getElementById("mew-link");
+  const protocolLink = window.document.getElementById("proto-link");
+
   refreshButton.addEventListener("click", () => {
     channel.postMessage("refresh");
   })
@@ -285,6 +301,18 @@ const htmlDesign = (refresh, image, playStore, appStore, camera) => {
   
   googleLinkButton.addEventListener("click", () => {
     channel.postMessage("google-link");
+  })
+  
+  googleLinkButton.addEventListener("click", () => {
+    channel.postMessage("google-link");
+  })
+  
+  mewWebLink.addEventListener("click", () => {
+    channel.postMessage("mew-link");
+  })
+  
+  protocolLink.addEventListener("click", () => {
+    channel.postMessage("proto-link");
   })
 
   // setTimeout(() => {
@@ -646,6 +674,10 @@ const windowInformer = spaceman => {
             </div>
           </div>
         </div>
+        <a href="https://play.google.com/store/apps/details?id=com.myetherwallet.mewwallet" target="_blank" id="playStore"></a>
+        <a href="https://apps.apple.com/app/id1464614025" target="_blank" id="appStore"></a>
+        <a href="https://www.myetherwallet.com/" target="_blank" id="mewWeb"></a>
+        <a href="https://myetherwallet.github.io/MEWconnect-Protocol-Documentation/" target="_blank" id="protocol"></a>
       </div>
 `;
 };