+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+### Release v2.1.1-RC.4.14
+- Change mechanism for opening app/play store links
+- Add popup link to MEW web and MEWconnect Protocol
+
+
+### Release v2.1.1-RC.4.13
+- Fix notice to unblock popups for app/play store links 4
+
+### Release v2.1.1-RC.4.12
+- Fix notice to unblock popups for app/play store links 3
+
+### Release v2.1.1-RC.4.11
+- Fix notice to unblock popups for app/play store links 2
+
+### Release v2.1.1-RC.4.10
+- Add logs to investigate link not working
+- Fix notice to unblock popups for app/play store links
+
+### Release v2.1.1-RC.4.9
+- Add logs to investigate link not working
+- Add notice to unblock popups for app/play store links
+
+### Release v2.1.1-RC.4.8
+- Add logs to investigate link not working
+
+### Release v2.1.1-RC.4.7
+- Fix X used for closing notifications
+- Add logs to investigate link not working
+
 ### Release v2.1.1-RC.4.6
 - Identify approval transactions
 - Use double logical not to check for infuraId